info-service/requirements.txt
django
djangorestframework
django-cors-headers
psycopg2-binary
python-dotenv
mock
coverage
django-storages[s3]
requests
